Here is a bit of C trivia: 1 - The functions isspace(), isalnum(), etc in ctype.h for some reason take an int as an argument. 2 - They are often implemented as macros that loko up truth values in an array. 3 - Thus, on picky compilers, the perfectly natural construct { char c; return isalnum(c) } fails with a strange error message about using a char as an array index. This fix casts all arguments to ctype functions to int . Ugly, but required.
git-svn-id: file:///afs/nd.edu/user37/ccl/svn/cctools/trunk@766 a4d8336d-3463-0410-8bba-c098c45d37a8

+++ abstractions/allpairs/src/allpairs_util.c
@@ -121,7 +121,7 @@ int file_line_count(const char *filename)
 
 	while(fgets(buffer, MAX_FILENAME_LEN, fp) != NULL) {
 		for(i = 0; i < strlen(buffer); i++) {
-			if(isspace(buffer[i]) != 1) {
+			if(isspace((int)buffer[i]) != 1) {
 				count++;
 				break;
 			}
